refactor: swiper context 文件名修改
